tested version of smc implementation

+++ spf/discrete_hmm_model.hpp
@@ -19,13 +19,13 @@ using namespace std;
 
 class DiscreteHMM : public ProblemSpecification<int>
 {
-	unsigned int num_states;
+	unsigned long num_states;
 	vector<double> initial_distn;
 	vector<vector<double> > transition_probs;
     vector<vector<double> > emission_probs;
     vector<int> obs;
 public:
-	DiscreteHMM(unsigned int num_states, vector<double> mu, vector<vector<double> > transition_probs, vector<vector<double> > emission_probs, vector<int> obs);
+	DiscreteHMM(unsigned long num_states, vector<double> mu, vector<vector<double> > transition_probs, vector<vector<double> > emission_probs, vector<int> obs);
 	unsigned long num_iterations();
 	std::pair<int, double> propose_initial(gsl_rng *random);
 	std::pair<int, double> propose_next(gsl_rng *random, int t, int curr);
